requirements.txt
streamlit>=1.28.0
torch>=2.0.0
torchvision>=0.15.0
opencv-python-headless>=4.8.0
numpy>=1.24.0,<2.0.0
plotly>=5.15.0
Pillow>=10.0.0
fpdf2>=2.7.0
supabase>=2.0.0
